# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Overthrone game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List available tasks
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a specific task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}/submit:
    post:
      operationId: submitTaskAnswer
      tags: [tasks]
      summary: Submit an answer for a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTaskBody"
      responses:
        "200":
          description: Submission result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskSubmissionResult"
        "400":
          description: Bad request
  /cards/task:
    post:
      operationId: useTaskCard
      tags: [cards]
      summary: Use Task Card to start solving a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UseTaskCardBody"
      responses:
        "200":
          description: Task card used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardActionResult"
        "400":
          description: Bad request
  /cards/attack:
    post:
      operationId: useAttackCard
      tags: [cards]
      summary: Use Attack Card to attack another team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UseAttackCardBody"
      responses:
        "200":
          description: Attack result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttackResult"
        "400":
          description: Bad request
  /cards/alliance:
    post:
      operationId: useAllianceCard
      tags: [cards]
      summary: Use Alliance Card to form/request an alliance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UseAllianceCardBody"
      responses:
        "200":
          description: Alliance result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardActionResult"
        "400":
          description: Bad request
  /cards/alliance/respond:
    post:
      operationId: respondToAlliance
      tags: [cards]
      summary: Accept or reject an alliance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AllianceResponseBody"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardActionResult"
  /cards/backstab:
    post:
      operationId: useBackstabCard
      tags: [cards]
      summary: Use Backstab Card against alliance partner
      responses:
        "200":
          description: Backstab result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardActionResult"
        "400":
          description: Bad request
  /cards/suspicion:
    post:
      operationId: useSuspicionCard
      tags: [cards]
      summary: Use Suspicion Card to accuse ally of backstab
      responses:
        "200":
          description: Suspicion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuspicionResult"
        "400":
          description: Bad request
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        hp:
          type: integer
        ap:
          type: integer
        members:
          type: array
          items:
            type: string
        isEliminated:
          type: boolean
        allianceId:
          type: ["integer", "null"]
        activeTaskId:
          type: ["integer", "null"]
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - hp
        - ap
        - members
        - isEliminated
        - allianceId
        - activeTaskId
        - isAdmin
        - createdAt
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [sudoku, math, ctf, algorithm]
        difficulty:
          type: string
          enum: [easy, medium, hard]
        apReward:
          type: integer
        content:
          type: string
        isActive:
          type: boolean
      required:
        - id
        - title
        - description
        - type
        - difficulty
        - apReward
        - content
        - isActive
    SubmitTaskBody:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer
    TaskSubmissionResult:
      type: object
      properties:
        correct:
          type: boolean
        apEarned:
          type: integer
        message:
          type: string
      required:
        - correct
        - apEarned
        - message
    UseTaskCardBody:
      type: object
      properties:
        taskId:
          type: integer
      required:
        - taskId
    UseAttackCardBody:
      type: object
      properties:
        targetTeamId:
          type: integer
        apSpent:
          type: integer
        useAI:
          type: boolean
      required:
        - targetTeamId
        - apSpent
    UseAllianceCardBody:
      type: object
      properties:
        targetTeamId:
          type: integer
      required:
        - targetTeamId
    AllianceResponseBody:
      type: object
      properties:
        requestingTeamId:
          type: integer
        accept:
          type: boolean
      required:
        - requestingTeamId
        - accept
    CardActionResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        teamState:
          $ref: "#/components/schemas/Team"
      required:
        - success
        - message
    AttackResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        damageDealt:
          type: integer
        apSpent:
          type: integer
        targetTeamId:
          type: integer
        targetTeamName:
          type: string
      required:
        - success
        - message
        - damageDealt
        - apSpent
        - targetTeamId
        - targetTeamName
    SuspicionResult:
      type: object
      properties:
        correct:
          type: boolean
        message:
          type: string
        accusingTeamEliminated:
          type: boolean
        accusedTeamEliminated:
          type: boolean
      required:
        - correct
        - message
        - accusingTeamEliminated
        - accusedTeamEliminated
...